Store image index in skopeo format
Skopeo doesn't store an original image index, instead it stores custom
index that includes reference only to the platform specific manifests.
This change stores the `index.json` in the format that `skopeo` expects.
As result, the app pulled by `composectl` is stored in the way that
makes `skopeo` happy, so it can be used to inject images to the docker.
Basically, this is needed for forward compatibility - an app is pulled
by the new utility while the aklite/aklite-apps/aklite-offline of the
older version and still uses `skopeo`.

Signed-off-by: Mike Sul <mike.sul@foundries.io>