ps: Properly find services based on the same image
Fix finding app services and checking whether they are running if there
are services based on the same image.

The algorithm should check/compare the service hash for each running
container that the dockerd returns, instead of adding the not running
service to the app service list with "config hash mismatch" error.

Signed-off-by: Mike Sul <mike.sul@foundries.io>