ps: Fix app running status checking if multiple versions are stored
There is no need to determine which of the stored app versions should be
checked for the running status if an app URI is specified as the `ps`
command parameter since the URI uniquely identifies the app version.

If `ps` is invoked without parameters then gather and print information
about running apps regardless whether they are present in the app store
or not. If there are multiple app versions in the store then just
running app URI is not determined, it is equal to an empty value.

Signed-off-by: Mike Sul <mike.sul@foundries.io>