check: Check "familiar" ref in addition to cannonical
If images are hosted in docker.io registry then they can be referenced
by a short (aka "familiar") URI format.
For example, "docker.io/library/nginx:stable-alpine" or
"docker.io/homeassistant/home-assistant:stable" can be referenced as
"nginx:stable-alpine" and  "homeassistant/home-assistant:stable"
respectively.

The dockerd removes the "docker.io" or "docker.io/library" prefix when
stores images in its store and as result returns the image URIs in
the shorten format in a response to "http://localhost/images/json?all=1"
request. For example, "nginx@sha256:<hash>" or
"homeassistant/home-assistant:<hash>".

The given change checks if the "familiar"/shorten image URI is present
in the image map returned by dockerd in addition to checking the regular
image URI as it is specified in a compose project and the tagged image
URI.

Signed-off-by: Mike Sul <mike.sul@foundries.io>