test: Run smoke test against vanilla registry
Extend the existing smoke test so it runs against two types of registry,
the foundries one (with patches to support custom OCI manifest) and the
vanilla registry.

The smoke test obviously fails for the vanilla registry because it
rejects the non OCI-compliant manifest of compose app generated by
`compose publish` command.

The commits following this one will fix the failure.

Signed-off-by: Mike Sul <mike.sul@foundries.io>