publish: Add ability to turn of layers manifest
Add a new parameter to the publish command to turn on/off app layers
manifest creation, posting, and inclusion its reference into the
manifest of app being published.

The layers manifest reference in a compose app manifest makes it non-OCI
compliant and the vanilla registry rejects requests to post such the
manifest.

The new parameter allows turning off app layers manifest addition to the
app manifest. As a result of it, a compose app manifest becomes OCI
compliant so it can be pushed to the vanilla registry which is proved by
the smoke test.

The app layers manifest is redundant in case of `composectl` usage for
app pulling since this utility retrieves meta about missing blobs from
registry and it knows sizes of missing blobs (archived one). The app
layers manifest is used by the older non composectl-based aklite
versions to avoid writing C++ code for communicating with registries and
retrieving blob metadata.

Signed-off-by: Mike Sul <mike.sul@foundries.io>