publish: Make layers manifest more self-sufficient
Minimize dependencies on the `distribution` code in the code that
generates the non-OCI compliant manifest containing metadata about all
app layers.
For some reason, the manifest type was defined as an image index while
it is actually is an image manifest since it contains layers.
We cannot change the manifest media format because older versions of
aklite depends on it. And the newer version of `distribution` code
checks consistency of the format. So, this code actually fails if the
newer version of `distribution` is used. This change removes dependency
on `distribution` hence no failure occurs even if a newer version of
`distribution` is used.

Signed-off-by: Mike Sul <mike.sul@foundries.io>