check: Unify app installation check
Use the same function to check whether app in installed for both the
"check" and "ps" commands. This check includes:
1. Verification if all app images are present in the docker store.
2. Verification if an app bundle is properly installed in the
   compose/project directory (all bundle files present and their hashes
   match the expected hashes).

Also, a new flag "--install" was added to the "ps" command to turn
on/off the installation check after checking if app is running.
By default this check is enabled.

Signed-off-by: Mike Sul <mike.sul@foundries.io>