uninstall: Add flag to optionally prune images
Add a new flag to enable/disable app images pruning in the docker store
just after removing an app installation from the compose/project
directory.

By default the flag is set to false (no pruning). If set to true, then
app images that are not used by any containers (regardless of a
container state) are removed from the docker store.

Signed-off-by: Mike Sul <mike.sul@foundries.io>