app: Skip app integrity check if pulled by skopeo
If app and its images are pulled by `skopeo` then we should not expect
the app bundle index in a local blob store even if the app manifest
includes a reference to the app bundle index.
The app bundle index presence in a local blob store is mandatory only if
an app is pulled by `composectl`.

Signed-off-by: Mike Sul <mike.sul@foundries.io>