check: Don't require all blobs if pulled by aklite+skopeo
If an app is pulled by "aklite+skopeo" (e.g. aklite v.92) then not all
app blobs are fetched to a blob store. Some of them are fetched and
stored in a app store and some of them are not fetched at all.

This change adjust `composectl check --local` and `composectl install`
commands so they does not require all blobs in a blob store if an app
is pulled by "aklite+skopeo". Instead, it requires all blobs needed to
build an app's Merkle tree, install and run apps.

Signed-off-by: Mike Sul <mike.sul@foundries.io>