app: Make app index optional if checked locally
If check whether app is fetched and installed is performed
locally (with `--local` option) and an app index is missing in the store
then do not include it into an app tree during building it. The reason
for is a fact that app can be pulled by `composectl` version or by
some other utility that is not aware about an app index, so it is never
pulled. In this case, we make an app index and it utilization optional,
so no network connectivity is required if any operation over an app is
conducted, e.g. `install`, `run`, etc.

An app index is required only if it is not a "local" check as well as
during app pulling - hence we guarantee that app index is pulled and
applied if a device is online and app is fully synced.

Signed-off-by: Mike Sul <mike.sul@foundries.io>