app_store: Avoid app bundle removal if blob store is pruned
Use hard links instead of symlinks for an app bundle archive and app
manifest in the app directory of an overall app store.

Using symlinks leads to removal of these two artifacts if downgrade or
rollback to LmP version <=93 occurs. The previous aklite versions do not
expect these two blobs in the blob store, they expect them in the app
directory of the store, so the prune operation removes them from the
blob store, and, as the result, the symlinks defined in the app
directory become invalid. It makes aklite thinks that app is not fully
fetched and starts the "sync" update which never succeeds, so aklite
ends up in doing eternal "sync" updates.

Replacing the symlinks with hard links fixes the issue since the blobs
are not removed because they have two links/references, one in the blobs
part and the second one in the app directory. Removing one of these hard
links does not lead to the blob removal.

Signed-off-by: Mike Sul <mike.sul@foundries.io>