uninstall: Remove app dir only if no other app version in store
Remove the app compose dir only if the app being removed is the only app
version in the store. If there are more than one version of the same app
in the store then we cannot remove its compose dir because this dir is
shared across many app versions (something to be improved in the
future).

Signed-off-by: Mike Sul <mike.sul@foundries.io>