provider: Use specialized remote blob reader for fetching blobs
Introduce and use a ReadCloser in the remote blob provider that takes a
Descriptor as input and skips digest calculation and verification during
reads.

This is suited for workflows like containerd's ingester, which performs
its own digest computation and verification while writing blob data.
Using this reader avoids redundant checks and improves efficiency.

Signed-off-by: Mike Sul <mike.sul@foundries.io>