fetch: Update read stat after fetch completion
Update the read statistics (average and current) at the exit of the read
monitor if new read stat data arrived after the last stat calculation.

Small blobs can be downloaded faster then sleep time of the read monitor
ticker. If so, the ticker timer is never fired so the stat calculation
is never performed and `0 B/s` is set as an average and current speed.
Therefore, adding stat re-calculation at the read monitor exit fixes
this issue.

Also, it makes the read stat value more precise if blob is downloaded
within a few read monitor ticks.

Signed-off-by: Mike Sul <mike.sul@foundries.io>