config: Fix app store creation
The app store factory method is supposed to use the current config
values as input parameters not the values set at the moment when the
factory method is defined. Otherwise, if the config values are
overridden, e.g. `Config.StoreRoot` then the factory method won't use
them and apply the previous one that were set before the override.

Signed-off-by: Mike Sul <mike.sul@foundries.io>