fetch: Ensure read monitor receives stop signal
If blob copying completes faster than the read monitor starts its
goroutine and listens for a stop signal at a stop channel, then a client
code invokes `readMonitor.Stop()` and send a signal through a
non-buffered channel while no one listens for it. As result, the stop
signal is never received by the read monitor goroutine and it hangs
forever.
There are two ways to fix it:
1. Making the call that sends the stop signal blocking.
2. Making the stop channel buffered channel and buffering exactly one
   value.

The given change does #2 as it is more reliable. The option #1 may still
lead to blocking if the read monitor exits before `Stop` is called.

Signed-off-by: Mike Sul <mike.sul@foundries.io>