fetch: Update blob state passed as input
The `FetchBlobs` caller passes a map of the blobs to be fetched via
pointer/reference and expects that the specified blob structures are
updated while corresponding blobs are being fetched.
Before this change the passed blob structures/info were set by value in
the internal structure of the blob fetching function. This change makes
sure that the input blob structures are set by reference hence any blob
state changes are visible to the caller.

Signed-off-by: Mike Sul <mike.sul@foundries.io>