ps: Infer app service health from its status
If a health check is not enabled in an app service then the "health"
field is not set in its status. This change infers an app service health
based on its status. If a service is running or created or exited with
zero exit code (one shot container/service), then consider it healthy.

Also, this change makes the app status reporting to Device Gateway in
sync with the previous versions of aklite, the versions that does not
use `composectl` for app management. In other words, it makes it
consistent with the previous versions of aklite.

Signed-off-by: Mike Sul <mike.sul@foundries.io>