status_printer: Take into account already fetched blobs
During update fetching it can happen that some of small blobs are
already fetched by the time the progress handler is invoked. If it
occurs then the progress printer should take into account the
current&average speed of such blobs. Otherwise, the speed values are
always zero if blobs are downloading between two subsequent calls of the
progress handler.

Signed-off-by: Mike Sul <mike.sul@foundries.io>