update: Don't do fetch if nothing to fetch
If moving an update state to "fetched" or "installed" at the update
initialization is turned off then it can happen that there is 0 bytes to
fetch when a user invokes the fetch command. In such cases it does not
make sense to start the app blobs fetching procedure at all.

Signed-off-by: Mike Sul <mike.sul@foundries.io>