stop: Skip stopping non-installed apps
If a user specifies a list of apps to stop and some of the apps are
running while some of them are not installed at all, then just ignore
non-installed apps and proceed with stopping installed apps.
It allows to stop apps even if some of them are not installed at all.
Otherwise, a client must make sure that all specified apps are installed
to stop them which impose additional burden on the client.

Signed-off-by: Mike Sul <mike.sul@foundries.io>