provider: Introduce support for satellite server
The satellite server, https://github.com/foundriesio/dg-satellite, needs
a way for devices to pick up registry content from it instead of
hub.foundries.io. This change introduces a mechanism to tell our
provider code to proxy its requests through the satellite server.

The error handling in this is a little awkward but allowed us to bubble
up errors w/o having to adjust a lot of users of the API.

Signed-off-by: Andy Doan <andy@foundries.io>