update: Don't fail update if its completion fails
Don't mark an update as "failed" if the "complete" fails.
The "complete" call is allowed only if an update is successfully started
or is in "completing" state, hence it doesn't make sense to "fail" it
because the update is actually successful.
Instead, if an error occurs, keep the update record in "completing"
state. A user/caller can call the forced completion if it is needed.

Signed-off-by: Mike Sul <mike.sul@foundries.io>