image_loader: Correctly count image loading responses from dockerd
Dockerd emits an "image loaded" response for each tag reference listed in
an image manifest. This means that if an image has N tag references,
dockerd will send the "image loaded" response N times, even though the
image blobs are only loaded once.

Additionally, when a digest reference is used in the "RepoTags" field of
an image load manifest, dockerd sends the "image loaded" response only
once, regardless of the number of tag references.

This commit updates the image loading logic to account for these
behaviors and correctly count the responses.

Signed-off-by: Mike Sul <mike.sul@foundries.io>