image_loader: Skip warning if expected index to exceed the max
In some cases it is expected that the image index becomes higher than
the maximum allowed index. In particular, when the last image is loaded
or when the last image is already present in the docker store.
The only case when this is a real error is if a client starts receiving
messages about loading a new image while the image index is already
equal to an overall image count.

Signed-off-by: Mike Sul <mike.sul@foundries.io>