update_test: Add test to verify app pruning
Ensure uninstall with pruning does not remove needed images

The test verifies that uninstalling apps with pruning does not remove
too many images, which could affect the detection of the currently
running apps.

Specifically, pruning all images without containers can remove tags
for images that are still used by running containers. This can lead
to a client incorrectly determining that an app is not running due
to a missing image.

Also, it seems like the issue in the docker since it removes wrong tags.

Signed-off-by: Mike Sul <mike.sul@foundries.io>