config: Define proxy providing function in config
Defining a compose app proxy as a function in the configuration
provides the compose app API clients to define runtime logic for
obtaining URL and/or CA cert of the proxy server.

In particular, a client does not need to know which of the API functions
makes a call to a container registry, hence it does not need to set the
proxy before invoking each such the API call. Instead composeapp invokes
a client provided function/callback to provide the proxy URL and CA
whenever a random API function of composeapp needs it.

Signed-off-by: Mike Sul <mike.sul@foundries.io>